pyproject.toml
[project]
name = "game-bot"
version = "0.2.0"
description = "Mobile game automation bot via ADB"
requires-python = ">=3.10"
dependencies = [
    "adbutils>=2.7.0",
    "opencv-python-headless>=4.9.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "schedule>=1.2.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]
web = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "websockets>=12.0",
]

[tool.ruff]
line-length = 120
